+++ src/Transformer/Implementation/NullTransformer.php
@@ -15,6 +15,7 @@ namespace Rekalogika\Mapper\Transformer\Implementation;
 
 use Rekalogika\Mapper\Context\Context;
 use Rekalogika\Mapper\Exception\InvalidArgumentException;
+use Rekalogika\Mapper\Transformer\Exception\NullSourceButMandatoryTargetException;
 use Rekalogika\Mapper\Transformer\TransformerInterface;
 use Rekalogika\Mapper\Transformer\TypeMapping;
 use Rekalogika\Mapper\Util\TypeCheck;
@@ -31,6 +32,21 @@ final readonly class NullTransformer implements TransformerInterface
         ?Type $targetType,
         Context $context,
     ): mixed {
+        // if the source is null & the target is an object, ignore it by
+        // returning the existing target. if the existing target is not already
+        // an object, throw an exception
+
+        if (TypeCheck::isObject($targetType)) {
+            if ($target === null) {
+                throw new NullSourceButMandatoryTargetException(
+                    targetType: $targetType,
+                    context: $context,
+                );
+            }
+
+            return $target;
+        }
+
         if ($target !== null) {
             throw new InvalidArgumentException('Target must be null');
         }
@@ -70,6 +86,7 @@ final readonly class NullTransformer implements TransformerInterface
         yield new TypeMapping(TypeFactory::null(), TypeFactory::float());
         yield new TypeMapping(TypeFactory::null(), TypeFactory::bool());
         yield new TypeMapping(TypeFactory::null(), TypeFactory::array());
+        yield new TypeMapping(TypeFactory::null(), TypeFactory::object(), true);
         yield new TypeMapping(TypeFactory::mixed(), TypeFactory::null());
     }
 }